pyproject.toml
[tool.poetry]
name = "meetmemo"
version = "0.1.0"
description = ""
authors = ["Leandro S. Silva"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.31.1"
openai = "^1.12.0"
python-dotenv = "^1.0.1"
streamlit-webrtc = "^0.47.1"
pydub = "^0.25.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
